+++ src/QuickMarcEditor/utils.js
@@ -300,6 +300,12 @@ const validateMarcBibRecord = (marcRecords) => {
 };
 
 const validateMarcHoldingsRecord = (marcRecords) => {
+  const instanceHridRecords = marcRecords.filter(({ tag }) => tag === '004');
+
+  if (instanceHridRecords.length > 1) {
+    return <FormattedMessage id="ui-quick-marc.record.error.instanceHrid.multiple" />;
+  }
+
   const locationRecords = marcRecords.filter(({ tag }) => tag === '852');
 
   if (!locationRecords.length) {