+++ src/QuickMarcEditor/utils.test.js
@@ -308,6 +308,23 @@ describe('QuickMarcEditor utils', () => {
 
         expect(utils.validateMarcRecord(record, initialValues, MARC_TYPES.HOLDINGS).props.id).toBe('ui-quick-marc.record.error.location.empty');
       });
+
+      it('should return error message when record has several 004 rows', () => {
+        const initialValues = { records: [] };
+        const record = {
+          leader: '04706cxm a22008651i 4500',
+          records: [
+            {
+              content: '04706cxm a22008651i 4500',
+              tag: 'LDR',
+            },
+            { tag: '004' },
+            { tag: '004' },
+          ],
+        };
+
+        expect(utils.validateMarcRecord(record, initialValues, MARC_TYPES.HOLDINGS).props.id).toBe('ui-quick-marc.record.error.instanceHrid.multiple');
+      });
     });
 
     it('should return error message when record has several 852 rows', () => {